refactor(providers): extract shared NDJSON line feed for claude/copilot parsers
- one canonical feed/flush split loop in _engine/line-feed.ts (capped via createCappedAppend)
- parsers keep only their per-line emit + state reduction
