+++ src/integration/ai/providers/copilot/parse-stream.ts
@@ -15,7 +15,7 @@ import type {
   CopilotStreamParser,
   CopilotUsage,
 } from '@src/integration/ai/providers/_engine/copilot-stream.ts';
-import { createCappedAppend } from '@src/integration/ai/providers/_engine/bounded-tail.ts';
+import { createCappedLineFeed } from '@src/integration/ai/providers/_engine/line-feed.ts';
 import { isRecord, numberField, stringField } from '@src/integration/ai/providers/_engine/json-field.ts';
 
 const extractUsage = (json: Record<string, unknown>): CopilotUsage | undefined => {
@@ -91,13 +91,6 @@ const extractBodyText = (json: Record<string, unknown>): string | undefined => {
 };
 
 export const createCopilotStreamParser = (): CopilotStreamParser => {
-  let buffer = '';
-  // Cap the in-flight line accumulator. A single NDJSON record embedding a large file-read /
-  // bash tool result can grow `buffer` to tens of MB before its newline clears it — an OOM-class
-  // accumulation. `feed` is the SOLE append site, so capping here keeps the invariant for `flush`
-  // too (it only drains an already-bounded buffer). Shared impl — see `createCappedAppend`
-  // (drop-oldest, one-shot warn).
-  const appendCapped = createCappedAppend('copilot-stream');
   const emit = (raw: string, onLine: (line: CopilotStreamLine) => void): void => {
     if (raw.length === 0) return;
     if (raw.startsWith('{') && raw.endsWith('}')) {
@@ -126,22 +119,9 @@ export const createCopilotStreamParser = (): CopilotStreamParser => {
     }
     onLine({ raw });
   };
+  const lineFeed = createCappedLineFeed<CopilotStreamLine>('copilot-stream', emit);
   return {
-    feed(chunk, onLine) {
-      buffer = appendCapped(buffer, chunk);
-      let nl = buffer.indexOf('\n');
-      while (nl !== -1) {
-        const line = buffer.slice(0, nl);
-        buffer = buffer.slice(nl + 1);
-        emit(line, onLine);
-        nl = buffer.indexOf('\n');
-      }
-    },
-    flush(onLine) {
-      if (buffer.length > 0) {
-        emit(buffer, onLine);
-        buffer = '';
-      }
-    },
+    feed: lineFeed.feed,
+    flush: lineFeed.flush,
   };
 };
